requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==11.3.0

openai>=1.0
